requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--extra-index-url http://pypi.buzzvil-internal.com
--trusted-host pypi.buzzvil-internal.com

adrecommender==0.1.5
airflow-exporter==1.3.0
alembic==1.2.1            # via apache-airflow
apache-airflow-backport-providers-amazon==2020.6.24
apache-airflow-backport-providers-mysql==2020.6.24
apache-airflow[kubernetes]==1.10.11
apispec[yaml]==1.3.3      # via flask-appbuilder
argcomplete==1.11.0       # via apache-airflow
arrow==0.12.1
attrs==19.3.0             # via apache-airflow, cattrs, jsonschema, pytest
awscli==1.18.114
babel==2.6.0              # via flask-babel
boto3==1.14.36
botocore==1.17.37         # via awscli, boto3, s3transfer
buzzvil_profile==0.1.10
cached-property==1.5.1    # via apache-airflow
cachetools==4.0.0         # via google-auth
cattrs==1.0.0             # via apache-airflow
certifi==2018.11.29       # via kubernetes, requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via flask, flask-appbuilder, pip-tools
colorama==0.3.9           # via awscli, flask-appbuilder
colorlog==4.0.2           # via apache-airflow
configparser==3.5.0       # via apache-airflow
croniter==0.3.28          # via apache-airflow
cryptography==2.8         # via apache-airflow
defusedxml==0.5.0         # via python3-openid
dill==0.2.8.2             # via apache-airflow
dnspython==2.0.0          # via email-validator
docutils==0.14            # via awscli, botocore, python-daemon
email-validator==1.1.1    # via apache-airflow
fastparquet==0.3.2
first==2.0.1              # via pip-tools
flake8==3.6.0
flask-admin==1.5.4        # via apache-airflow
flask-appbuilder==2.2.2   # via apache-airflow
flask-babel==0.12.2       # via flask-appbuilder
flask-caching==1.3.3      # via apache-airflow
flask-jwt-extended==3.24.0  # via flask-appbuilder
flask-login==0.4.1        # via apache-airflow, flask-appbuilder
flask-openid==1.2.5       # via flask-appbuilder
flask-sqlalchemy==2.4.1   # via flask-appbuilder
flask-swagger==0.2.13     # via apache-airflow
flask-wtf==0.14.2         # via apache-airflow, flask-appbuilder
flask==1.1.1              # via apache-airflow, flask-admin, flask-appbuilder, flask-babel, flask-caching, flask-jwt-extended, flask-login, flask-openid, flask-sqlalchemy, flask-swagger, flask-wtf
funcsigs==1.0.0           # via apache-airflow
future==0.16.0            # via apache-airflow
google-auth==1.10.0
googleapis-common-protos==1.6.0  # via adrecommender, buzzvil-profile
graphviz==0.13.2          # via apache-airflow
grpcio==1.26.0
gunicorn==19.9.0          # via apache-airflow
idna==2.8                 # via email-validator, requests
importlib-metadata==0.23  # via jsonschema
iso8601==0.1.12           # via apache-airflow
itsdangerous==1.1.0       # via flask
jinja2==2.10.3            # via apache-airflow, flask, flask-babel, python-nvd3
jmespath==0.9.3           # via boto3, botocore
json-merge-patch==0.2     # via apache-airflow
jsonschema==3.1.1         # via apache-airflow, flask-appbuilder
kubernetes==10.0.1        # via apache-airflow
lazy-object-proxy==1.4.2  # via apache-airflow
llvmlite==0.31.0          # via numba
lockfile==0.12.2          # via python-daemon
mako==1.0.7               # via alembic
markdown==2.6.11          # via apache-airflow
markupsafe==1.1.0         # via jinja2, mako
marshmallow-enum==1.5.1   # via flask-appbuilder
marshmallow-sqlalchemy==0.18.0  # via flask-appbuilder
marshmallow==2.19.5       # via flask-appbuilder, marshmallow-enum, marshmallow-sqlalchemy
mccabe==0.6.1             # via flake8
more-itertools==7.2.0     # via pytest, zipp
mysql-connector-python==8.0.18  # via apache-airflow-backport-providers-mysql
mysqlclient==1.3.14       # via apache-airflow-backport-providers-mysql
numba==0.48.0             # via fastparquet
numpy==1.16.0
oauthlib==3.1.0           # via requests-oauthlib
ordereddict==1.1          # via funcsigs
packaging==20.1           # via pytest
pandas==1.0.5
pendulum==1.4.4           # via apache-airflow
pip-tools==2.0.1
pluggy==0.13.1            # via pytest
prison==0.1.2             # via flask-appbuilder
prometheus-client==0.8.0  # via airflow-exporter
protobuf==3.11.2          # via googleapis-common-protos, mysql-connector-python
psutil==5.7.0
psycopg2==2.8.6
py==1.8.1                 # via pytest, pytest-datafiles
pyarrow==0.17.1
pyasn1-modules==0.2.7     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycodestyle==2.4.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.0.0           # via flake8
pygments==2.3.1           # via apache-airflow
pyjwt==1.7.1              # via flask-appbuilder, flask-jwt-extended
pyparsing==2.4.6          # via packaging
pyrsistent==0.15.4        # via jsonschema
pytest-datafiles==2.0
pytest-flake8==1.0.5
pytest-helpers-namespace==2019.1.8
pytest-mock==2.0.0
pytest==5.4.2
python-daemon==2.1.2      # via apache-airflow
python-dateutil==2.7.5    # via alembic, apache-airflow, arrow, botocore, croniter, flask-appbuilder, kubernetes, pandas, pendulum
python-editor==1.0.3      # via alembic
python-nvd3==0.15.0       # via apache-airflow
python-slugify==4.0.1     # via apache-airflow, python-nvd3
python3-openid==3.1.0     # via flask-openid
pytz==2018.7              # via babel, pandas, tzlocal
pytzdata==2018.7          # via pendulum
pyyaml==5.1.2
requests-oauthlib==1.3.0  # via kubernetes
requests==2.21.0          # via apache-airflow, kubernetes, requests-oauthlib
rsa==3.4.2                # via awscli, google-auth
s3transfer==0.3.3         # via awscli, boto3
setproctitle==1.1.10      # via apache-airflow
six==1.11.0               # via cryptography, fastparquet, flask-jwt-extended, google-auth, grpcio, jsonschema, kubernetes, packaging, pip-tools, prison, protobuf, pyrsistent, python-dateutil, sqlalchemy-utils, tenacity, thrift, websocket-client
sqlalchemy-jsonfield==0.9.0  # via apache-airflow
sqlalchemy-utils==0.36.1  # via flask-appbuilder
sqlalchemy==1.3.10        # via alembic, apache-airflow, flask-sqlalchemy, marshmallow-sqlalchemy, sqlalchemy-jsonfield, sqlalchemy-utils
tabulate==0.8.2           # via apache-airflow
tenacity==4.12.0          # via apache-airflow
text-unidecode==1.3       # via python-slugify
thrift==0.11.0            # via apache-airflow, fastparquet
tzlocal==1.5.1            # via apache-airflow, pendulum
unicodecsv==0.14.1        # via apache-airflow
urllib3==1.24.3
watchtower==0.7.3         # via apache-airflow-backport-providers-amazon
wcwidth==0.1.8            # via pytest
websocket-client==0.57.0  # via kubernetes
werkzeug==0.16.0          # via apache-airflow, flask, flask-caching, flask-jwt-extended
wtforms==2.2.1            # via flask-admin, flask-wtf
zipp==0.6.0               # via importlib-metadata
zope.deprecation==4.4.0   # via apache-airflow
